move component into new file
