+++ packages/metadata-common/src/index.ts
@@ -18,3 +18,4 @@ import '../style/index.css';
 
 export * from './MetadataEditor';
 export * from './MetadataWidget';
+export * from './AddMetadataButton';
